+++ packages/liveblocks-react-comments/src/primitives/EmojiPicker/types.ts
@@ -14,6 +14,8 @@ export type Emoji = {
   tags?: string[];
 };
 
+export type IndexedEmoji = Emoji & { index: number };
+
 export type EmojiCategory = {
   key: number;
   name: string;
@@ -30,9 +32,7 @@ export type EmojiData = {
   skinTones: EmojiSkinTone[];
 };
 
-export type EmojiPickerRow = {
-  emojis: Emoji[];
-};
+export type EmojiPickerRow = IndexedEmoji[];
 
 export type EmojiPickerData = {
   count: number;
@@ -42,9 +42,15 @@ export type EmojiPickerData = {
   categoriesRowIndices: number[][];
 };
 
+export type EmojiPickerSelectionDirection = "left" | "right" | "up" | "down";
+
+export type EmojiPickerInteraction = "keyboard" | "pointer" | "none";
+
 export type EmojiPickerContentLoadingProps = ComponentPropsWithoutRef<"div">;
 
 export type EmojiPickerContentEmptyProps = ComponentPropsWithoutRef<"div">;
+
+export type EmojiPickerContentGridProps = ComponentPropsWithoutRef<"div">;
 
 export interface EmojiPickerContentErrorProps
   extends ComponentPropsWithoutRef<"div"> {
@@ -79,7 +85,7 @@ export type EmojiPickerContentEmojiRowAttributes = {
   categoryRowsCount: number;
 };
 
-export interface EmojiPickerContentEmojiRowProps
+export interface EmojiPickerContentRowProps
   extends ComponentPropsWithoutRef<"div"> {
   /**
    * Attributes related to the current row.
@@ -134,9 +140,14 @@ export interface EmojiPickerContentComponents {
   CategoryHeader: ComponentType<EmojiPickerContentCategoryHeaderProps>;
 
   /**
-   * The component used to display emoji rows.
+   * The component displayed when there are emojis to display.
    */
-  EmojiRow: ComponentType<EmojiPickerContentEmojiRowProps>;
+  Grid: ComponentType<EmojiPickerContentGridProps>;
+
+  /**
+   * The component used to display rows of emojis.
+   */
+  Row: ComponentType<EmojiPickerContentRowProps>;
 
   /**
    * The component used to display emojis.